requirements.txt.txt
openai-agents
openai
pandas
numpy
scipy
plotly
streamlit
openpyxl
python-dotenv
reportlab